pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "arrow",
    "click",
    "css-inline",
    "glom",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "jinja2",
    "marko",
    "python-dotenv",
    "requests",
    "rich",
    "tzlocal",
]

[project.scripts]
convert-timezone = "iim.convert_timezone:convert_timezone"
iim-data = "iim.iim_data:iim_data"
iim-gdoc-download = "iim.iim_gdoc_download:iim_gdoc_download"
iim-gdocs-to-jira = "iim.iim_gdocs_to_jira:iim_google_docs_to_jira"
iim-lint = "iim.iim_lint:iim_lint"
iim-sync = "iim.iim_sync:iim_sync"
iim-to-review = "iim.iim_to_review:iim_to_review"
iim-qbr = "iim.iim_qbr:iim_qbr"
iim-weekly-overview = "iim.iim_weekly_overview:iim_weekly_report"

[dependency-groups]
dev = [
    "pytest",
    "responses",
    "ruff",
    "ty",
]
